As part of an effort to extend Book-E to the 64-bit world, make the necessary
changes to the DPAA/dTSEC driver set to be integer-pointer conversion clean.
This means no more casts to int, and use uintptr_t where needed.

Since the NCSW source is effectively obsolete, direct changes to the source tree
are safe.

+++ sys/dev/dpaa/bman.c
@@ -114,7 +114,7 @@ bman_attach(device_t dev)
 	bp.totalNumOfBuffers = BMAN_MAX_BUFFERS;
 	bp.f_Exception = bman_exception;
 	bp.h_App = sc;
-	bp.errIrq = (int)sc->sc_ires;
+	bp.errIrq = (uintptr_t)sc->sc_ires;
 	bp.partBpidBase = 0;
 	bp.partNumOfPools = BM_MAX_NUM_OF_POOLS;
 	printf("base address: %llx\n", (uint64_t)bp.baseAddress);
